Fix aliasing of results for duplicate keys in raven read requests
If the same key was requested multiple times in a single read request, the
result would include the same instance of the requested document multiple
times. Now a new instance of the document and its metadata is created for
each instance of the key. This fixes issues where one instance was
requested for read while the other was requested for modification.