Make sure we run the onComplete handler even if there was an exception
This refactors the MutableOrderedGroupingBoundedQueue enqueue function
slightly so that we can be sure we always trigger onComplete for every
item that was added. If we don't do this then a failed item results in the
projector position never being updated again (because that event is never
marked as complete in the projector's LastCompleteItemAgent).

Additionally, the try-with block now only wraps the user function (itemF).