Make unit builders distinct by name in AggregateStateBuilder
Currently the 'same' state builder is passed into AggregateStateBuilder
multiple times. It appears once per command / event handler that it is
used in.

This change filters out duplicates of state builders (by name). A warning
is also logged if the duplicates are different instances (this is just a
warning because they're probably only different instances because of being
passed through StateBuilder.withUnitKey).

Having duplicate state builders is an issue because of how the state
builders are run. Each state builder gets the output of the previous state
builder as its input, instead of getting the original state. This means
that ordering of state builders matters if two state builders modify the
same state. It also means that if duplicate state builders are present and
the state builders aren't idempotent they can produce invalid state (seems
that non-idempotent state builders should be allowed, e.g. a state builder
that counts occurences of an event by just incrementing the count).

To fix this properly I think state builders need to be split into two
parts. Currently we essentially have a ('TState -> 'TEvent -> 'TState)
list (ignoring metadata for the moment). The only way to calculate the
state change using this is by applying each sequentially. But this has the
problem of later state builders seeing the partially modified state, not
the original state. Instead we need something like a ('TState -> 'TEvent
-> 'TChange) list and a ('TChange list -> 'TState) merge function.