readStreamSliceForward now only returns empty array at end of stream
This allows us to correctly detect the end of the stream when events have
been deleted from the stream (e.g. due to MaxCount setting). Also modified
the code to deal with the fact that the first event in the slice may not
be the event that we tried to read.